Overwriting the SQLAlchemy test for foreign key options
The test needs to be overwritten, because in case of HANA, if a user tries to reflect the referential action for foreign keys, they are always reflected even if the user does not explicitly change the options to be used, i.e. the default referential action being used by the database is reflected. In no case in SQLAlchemy-hana an empty dictionary is returned for foreign key options.