Fixing get_unique_constraints, improve get_table_names, overwriting test for foreign key options (#44)
* Fixing get_unique_constraints

Also those unique constraints should be returned which do not have a user defined name, but primary keys should not be returned as unique constraints

* Move the filter to HANA

* Making the select statements more robust

* Small changes to make the dialect consistent

* Overwriting the SQLAlchemy test for foreign key options

The test needs to be overwritten, because in case of HANA, if a user tries to reflect the referential action for foreign keys, they are always reflected even if the user does not explicitly change the options to be used, i.e. the default referential action being used by the database is reflected. In no case in SQLAlchemy-hana an empty dictionary is returned for foreign key options.

* Mention the reason for overwriting SQLAlchemy tests